Refactor app into blueprints with factory
