+++ run.py
@@ -1,4 +1,10 @@
-from website.app import app
+"""Development entrypoint for the Pagina application."""
+
+from website import create_app
+
+
+app = create_app()
+
 
 if __name__ == "__main__":
     app.run(debug=True, host="127.0.0.1", port=5000)
