+++ tests/test_login_auth.py
@@ -7,15 +7,16 @@ import pytest
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 sys.modules.setdefault('website.scheduler', types.SimpleNamespace())
 
-import website.app as app_module
+from website import create_app
+import website.auth as auth_module
 
-app = app_module.app
-add_to_allowlist = app_module.add_to_allowlist
+app = create_app()
+add_to_allowlist = auth_module.add_to_allowlist
 
 
 @pytest.fixture(autouse=True)
 def temp_allowlist(tmp_path):
-    app_module.ALLOWLIST_FILE = tmp_path / "allowlist.json"
+    auth_module.ALLOWLIST_FILE = tmp_path / "allowlist.json"
     yield
