+++ website/routes/__init__.py
@@ -1,0 +1,6 @@
+"""Routes package containing the main application blueprint."""
+
+from .generator import generator_bp
+
+__all__ = ["generator_bp"]
+
