+++ website/templates/checkout.html
@@ -55,7 +55,7 @@
 
             <div class="d-flex align-items-center gap-2 small text-muted">
               <i class="bi bi-envelope" aria-hidden="true"></i>
-              ¿Dudas? <a class="ms-1" href="{{ url_for('contacto') }}">Contáctanos</a>
+              ¿Dudas? <a class="ms-1" href="{{ url_for('payments.contacto') }}">Contáctanos</a>
             </div>
           </div>
         </div>
