+++ website/templates/login.html
@@ -15,7 +15,7 @@
   <div class="auth-form">
     <div class="card auth-card p-4 shadow-soft">
       <h4 class="mb-3">Iniciar sesión</h4>
-      <form method="post" action="{{ url_for('login') }}">
+      <form method="post" action="{{ url_for('auth.login') }}">
         <div class="mb-3">
           <label class="form-label">Correo</label>
           <input name="email" type="email" class="form-control" required placeholder="tucorreo@ejemplo.com" value="{{ request.form.get('email','') }}">
@@ -27,7 +27,7 @@
         <button type="submit" class="btn btn-primary w-100">Entrar</button>
       </form>
       <div class="d-flex justify-content-center gap-2 small mt-3">
-        <a href="{{ url_for('login') }}">Login</a>
+        <a href="{{ url_for('auth.login') }}">Login</a>
       </div>
     </div>
   </div>
