+++ website/templates/subscribe.html
@@ -46,7 +46,7 @@
 
             <div class="d-flex align-items-center gap-2 small text-muted">
               <i class="bi bi-envelope" aria-hidden="true"></i>
-              ¿Dudas? <a class="ms-1" href="{{ url_for('contacto') }}">Contáctanos</a>
+              ¿Dudas? <a class="ms-1" href="{{ url_for('payments.contacto') }}">Contáctanos</a>
             </div>
           </div>
         </div>
@@ -71,7 +71,7 @@
           body: JSON.stringify({ subscription_id: sub })
         });
         if (!res.ok) throw new Error(await res.text());
-        window.location.href = "{{ url_for('subscribe_success') }}?sub=" + encodeURIComponent(sub);
+        window.location.href = "{{ url_for('payments.subscribe_success') }}?sub=" + encodeURIComponent(sub);
       } catch (err) {
         console.error(err);
         alert('Error con PayPal al activar.');
